Fix console-boot config when missing lightdm
Per #11, the lightdm check wrapping the handling for the B1 and B2 options
effectively turns them into a no-op when the user doesn't have lightdm installed
(quite likely for console-only and headless systems).  This would prevent users
from switching between the B1 (multi-user) and B2 (single-user) options.

This restores the original placement of the lightdm check from asb's original script
so that the sing/multi-user options will still be set.